assets/agent/requirements.txt
requests>=2.31.0
flask>=3.0.0
python-dateutil>=2.8.2
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-api>=1.20.0
a2a-sdk>=0.2.0
pyyaml>=6.0.0
python-json-logger>=3.3.0
sap-cloud-sdk @ https://common.repositories.cloud.sap/artifactory/api/pypi/application-foundation-sdk-python/sap-cloud-sdk/0.5.0.dev20260408+contribs.extensibility.e9a8908/sap_cloud_sdk-0.5.0.dev20260408+contribs.extensibility.e9a8908-py3-none-any.whl
